Implement redirection handling.
Still some issues. The new default behavior is to skip images if the
tool detectes a redirect link. If the CLI flask -a/--allow_redirects is
used the tool attempts to follow the redirection and render that
result. For some test cased I've run locally the redirect results are
often not as expected (one year redirects to another, ...) so I still
have some exploring to do.

Issue #5